LangChain-Intro/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bson==0.5.10
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.2
click==8.1.7
dnspython==2.7.0
fastapi==0.115.5
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langsmith==0.3.43
motor==3.6.0
orjson==3.10.18
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.41.3
tenacity==9.1.2
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.32.1
zstandard==0.23.0
